fully refactored, NewLayout renamed to Layout and javadocs regenerated
