+++ docs/allclasses-index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="en">
 <head>
-<!-- Generated by javadoc (17) on Sun Dec 10 09:26:44 GMT 2023 -->
+<!-- Generated by javadoc (17) on Sun Dec 10 16:29:36 GMT 2023 -->
 <title>All Classes and Interfaces</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
@@ -78,12 +78,14 @@ loadScripts(document, 'script');</script>
 </div>
 <div class="col-first even-row-color all-classes-table all-classes-table-tab2"><a href="com/github/westsi/browser/util/html/HTMLParser.html" title="class in com.github.westsi.browser.util.html">HTMLParser</a></div>
 <div class="col-last even-row-color all-classes-table all-classes-table-tab2">
-<div class="block">Lexes and parses source HTML into <code>HTMLTag</code>s and <code>HTMLText</code>.</div>
+<div class="block">Lexes and parses source HTML into a tree of <code>HTMLElement</code>s.</div>
 </div>
 <div class="col-first odd-row-color all-classes-table all-classes-table-tab2"><a href="com/github/westsi/browser/util/html/HTMLTag.html" title="class in com.github.westsi.browser.util.html">HTMLTag</a></div>
 <div class="col-last odd-row-color all-classes-table all-classes-table-tab2">&nbsp;</div>
 <div class="col-first even-row-color all-classes-table all-classes-table-tab2"><a href="com/github/westsi/browser/util/html/HTMLText.html" title="class in com.github.westsi.browser.util.html">HTMLText</a></div>
-<div class="col-last even-row-color all-classes-table all-classes-table-tab2">&nbsp;</div>
+<div class="col-last even-row-color all-classes-table all-classes-table-tab2">
+<div class="block">Represents a textual node in the DOM tree.</div>
+</div>
 <div class="col-first odd-row-color all-classes-table all-classes-table-tab2"><a href="com/github/westsi/browser/Layout.html" title="class in com.github.westsi.browser">Layout</a></div>
 <div class="col-last odd-row-color all-classes-table all-classes-table-tab2">
 <div class="block">Lays out a web page for a <code>BrowserTab</code>.</div>
