+++ docs/com/github/westsi/browser/Browser.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="en">
 <head>
-<!-- Generated by javadoc (17) on Sun Dec 10 09:26:44 GMT 2023 -->
+<!-- Generated by javadoc (17) on Sun Dec 10 16:29:36 GMT 2023 -->
 <title>Browser</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
