+++ docs/com/github/westsi/browser/Layout.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="en">
 <head>
-<!-- Generated by javadoc (17) on Sun Dec 10 09:26:44 GMT 2023 -->
+<!-- Generated by javadoc (17) on Sun Dec 10 16:29:36 GMT 2023 -->
 <title>Layout</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
@@ -112,14 +112,17 @@ loadScripts(document, 'script');</script>
 <div class="col-second even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code><a href="#getDisplayList()" class="member-name-link">getDisplayList</a>()</code></div>
 <div class="col-last even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4">&nbsp;</div>
 <div class="col-first odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code>void</code></div>
-<div class="col-second odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code><a href="#LayoutWebPage(java.util.ArrayList)" class="member-name-link">LayoutWebPage</a><wbr>(<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/util/ArrayList.html" title="class or interface in java.util" class="external-link">ArrayList</a>&lt;<a href="util/html/HTMLElement.html" title="interface in com.github.westsi.browser.util.html">HTMLElement</a>&gt;&nbsp;tokens)</code></div>
+<div class="col-second odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code><a href="#layout(com.github.westsi.browser.util.html.HTMLElement)" class="member-name-link">layout</a><wbr>(<a href="util/html/HTMLElement.html" title="interface in com.github.westsi.browser.util.html">HTMLElement</a>&nbsp;root)</code></div>
 <div class="col-last odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4">
 <div class="block">Lays out tokens on page, with handling for some <code>HTMLTag</code>s and <code>HTMLText</code>.</div>
 </div>
 <div class="col-first even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code>void</code></div>
-<div class="col-second even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code><a href="#updateResize(java.lang.Integer,java.lang.Integer)" class="member-name-link">updateResize</a><wbr>(<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Integer.html" title="class or interface in java.lang" class="external-link">Integer</a>&nbsp;width,
- <a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Integer.html" title="class or interface in java.lang" class="external-link">Integer</a>&nbsp;height)</code></div>
+<div class="col-second even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code><a href="#resetVariables()" class="member-name-link">resetVariables</a>()</code></div>
 <div class="col-last even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4">&nbsp;</div>
+<div class="col-first odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code>void</code></div>
+<div class="col-second odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code><a href="#updateResize(java.lang.Integer,java.lang.Integer)" class="member-name-link">updateResize</a><wbr>(<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Integer.html" title="class or interface in java.lang" class="external-link">Integer</a>&nbsp;width,
+ <a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Integer.html" title="class or interface in java.lang" class="external-link">Integer</a>&nbsp;height)</code></div>
+<div class="col-last odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4">&nbsp;</div>
 </div>
 </div>
 </div>
@@ -160,21 +163,27 @@ loadScripts(document, 'script');</script>
 </section>
 </li>
 <li>
-<section class="detail" id="LayoutWebPage(java.util.ArrayList)">
-<h3>LayoutWebPage</h3>
-<div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="return-type">void</span>&nbsp;<span class="element-name">LayoutWebPage</span><wbr><span class="parameters">(<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/util/ArrayList.html" title="class or interface in java.util" class="external-link">ArrayList</a>&lt;<a href="util/html/HTMLElement.html" title="interface in com.github.westsi.browser.util.html">HTMLElement</a>&gt;&nbsp;tokens)</span></div>
-<div class="block">Lays out tokens on page, with handling for some <code>HTMLTag</code>s and <code>HTMLText</code>.</div>
-<dl class="notes">
-<dt>Parameters:</dt>
-<dd><code>tokens</code> - The lexed <code>HTMLElement</code>s that need to be lain out.</dd>
-</dl>
-</section>
-</li>
-<li>
 <section class="detail" id="updateResize(java.lang.Integer,java.lang.Integer)">
 <h3>updateResize</h3>
 <div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="return-type">void</span>&nbsp;<span class="element-name">updateResize</span><wbr><span class="parameters">(<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Integer.html" title="class or interface in java.lang" class="external-link">Integer</a>&nbsp;width,
  <a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Integer.html" title="class or interface in java.lang" class="external-link">Integer</a>&nbsp;height)</span></div>
+</section>
+</li>
+<li>
+<section class="detail" id="resetVariables()">
+<h3>resetVariables</h3>
+<div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="return-type">void</span>&nbsp;<span class="element-name">resetVariables</span>()</div>
+</section>
+</li>
+<li>
+<section class="detail" id="layout(com.github.westsi.browser.util.html.HTMLElement)">
+<h3>layout</h3>
+<div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="return-type">void</span>&nbsp;<span class="element-name">layout</span><wbr><span class="parameters">(<a href="util/html/HTMLElement.html" title="interface in com.github.westsi.browser.util.html">HTMLElement</a>&nbsp;root)</span></div>
+<div class="block">Lays out tokens on page, with handling for some <code>HTMLTag</code>s and <code>HTMLText</code>.</div>
+<dl class="notes">
+<dt>Parameters:</dt>
+<dd><code>root</code> - The root of the <code>HTMLElement</code> tree that needs to be laid out.</dd>
+</dl>
 </section>
 </li>
 </ul>
