+++ docs/com/github/westsi/browser/util/html/package-summary.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="en">
 <head>
-<!-- Generated by javadoc (17) on Sun Dec 10 09:26:44 GMT 2023 -->
+<!-- Generated by javadoc (17) on Sun Dec 10 16:29:36 GMT 2023 -->
 <title>com.github.westsi.browser.util.html</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
@@ -92,12 +92,14 @@ loadScripts(document, 'script');</script>
 </div>
 <div class="col-first odd-row-color class-summary class-summary-tab2"><a href="HTMLParser.html" title="class in com.github.westsi.browser.util.html">HTMLParser</a></div>
 <div class="col-last odd-row-color class-summary class-summary-tab2">
-<div class="block">Lexes and parses source HTML into <code>HTMLTag</code>s and <code>HTMLText</code>.</div>
+<div class="block">Lexes and parses source HTML into a tree of <code>HTMLElement</code>s.</div>
 </div>
 <div class="col-first even-row-color class-summary class-summary-tab2"><a href="HTMLTag.html" title="class in com.github.westsi.browser.util.html">HTMLTag</a></div>
 <div class="col-last even-row-color class-summary class-summary-tab2">&nbsp;</div>
 <div class="col-first odd-row-color class-summary class-summary-tab2"><a href="HTMLText.html" title="class in com.github.westsi.browser.util.html">HTMLText</a></div>
-<div class="col-last odd-row-color class-summary class-summary-tab2">&nbsp;</div>
+<div class="col-last odd-row-color class-summary class-summary-tab2">
+<div class="block">Represents a textual node in the DOM tree.</div>
+</div>
 </div>
 </div>
 </div>
