+++ docs/index-files/index-1.html
@@ -1,12 +1,12 @@
 <!DOCTYPE HTML>
 <html lang="en">
 <head>
-<!-- Generated by javadoc (17) on Sun Dec 10 09:26:44 GMT 2023 -->
-<title>B-Index</title>
+<!-- Generated by javadoc (17) on Sun Dec 10 16:29:36 GMT 2023 -->
+<title>A-Index</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2023-12-10">
-<meta name="description" content="index: B">
+<meta name="description" content="index: A">
 <meta name="generator" content="javadoc/IndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.min.css" title="Style">
@@ -50,21 +50,19 @@ loadScripts(document, 'script');</script>
 <div class="header">
 <h1>Index</h1>
 </div>
-<a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">F</a>&nbsp;<a href="index-6.html">G</a>&nbsp;<a href="index-7.html">H</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">P</a>&nbsp;<a href="index-11.html">R</a>&nbsp;<a href="index-12.html">S</a>&nbsp;<a href="index-13.html">T</a>&nbsp;<a href="index-14.html">U</a>&nbsp;<a href="index-15.html">V</a>&nbsp;<a href="index-16.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a>
-<h2 class="title" id="I:B">B</h2>
+<a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">E</a>&nbsp;<a href="index-6.html">F</a>&nbsp;<a href="index-7.html">G</a>&nbsp;<a href="index-8.html">H</a>&nbsp;<a href="index-9.html">L</a>&nbsp;<a href="index-10.html">M</a>&nbsp;<a href="index-11.html">P</a>&nbsp;<a href="index-12.html">R</a>&nbsp;<a href="index-13.html">S</a>&nbsp;<a href="index-14.html">T</a>&nbsp;<a href="index-15.html">U</a>&nbsp;<a href="index-16.html">V</a>&nbsp;<a href="index-17.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a>
+<h2 class="title" id="I:A">A</h2>
 <dl class="index">
-<dt><a href="../com/github/westsi/browser/Browser.html" class="type-name-link" title="class in com.github.westsi.browser">Browser</a> - Class in <a href="../com/github/westsi/browser/package-summary.html">com.github.westsi.browser</a></dt>
-<dd>
-<div class="block">The overall browser window.&nbsp; Responsible for handling <code>BrowserTab</code>s and the global <code>JFrame</code>.</div>
-</dd>
-<dt><a href="../com/github/westsi/browser/BrowserTab.html" class="type-name-link" title="class in com.github.westsi.browser">BrowserTab</a> - Class in <a href="../com/github/westsi/browser/package-summary.html">com.github.westsi.browser</a></dt>
-<dd>
-<div class="block">Handles layout, rendering, and interactivity for an individual browser tab.</div>
-</dd>
-<dt><a href="../com/github/westsi/browser/BrowserTab.html#%3Cinit%3E(java.lang.String,java.lang.Integer,java.lang.Integer)" class="member-name-link">BrowserTab(String, Integer, Integer)</a> - Constructor for class com.github.westsi.browser.<a href="../com/github/westsi/browser/BrowserTab.html" title="class in com.github.westsi.browser">BrowserTab</a></dt>
+<dt><a href="../com/github/westsi/browser/util/html/HTMLTag.html#addAttribute(java.lang.String,java.lang.String)" class="member-name-link">addAttribute(String, String)</a> - Method in class com.github.westsi.browser.util.html.<a href="../com/github/westsi/browser/util/html/HTMLTag.html" title="class in com.github.westsi.browser.util.html">HTMLTag</a></dt>
+<dd>&nbsp;</dd>
+<dt><a href="../com/github/westsi/browser/util/html/HTMLElement.html#addChild(com.github.westsi.browser.util.html.HTMLElement)" class="member-name-link">addChild(HTMLElement)</a> - Method in interface com.github.westsi.browser.util.html.<a href="../com/github/westsi/browser/util/html/HTMLElement.html" title="interface in com.github.westsi.browser.util.html">HTMLElement</a></dt>
+<dd>&nbsp;</dd>
+<dt><a href="../com/github/westsi/browser/util/html/HTMLTag.html#addChild(com.github.westsi.browser.util.html.HTMLElement)" class="member-name-link">addChild(HTMLElement)</a> - Method in class com.github.westsi.browser.util.html.<a href="../com/github/westsi/browser/util/html/HTMLTag.html" title="class in com.github.westsi.browser.util.html">HTMLTag</a></dt>
+<dd>&nbsp;</dd>
+<dt><a href="../com/github/westsi/browser/util/html/HTMLText.html#addChild(com.github.westsi.browser.util.html.HTMLElement)" class="member-name-link">addChild(HTMLElement)</a> - Method in class com.github.westsi.browser.util.html.<a href="../com/github/westsi/browser/util/html/HTMLText.html" title="class in com.github.westsi.browser.util.html">HTMLText</a></dt>
 <dd>&nbsp;</dd>
 </dl>
-<a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">F</a>&nbsp;<a href="index-6.html">G</a>&nbsp;<a href="index-7.html">H</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">P</a>&nbsp;<a href="index-11.html">R</a>&nbsp;<a href="index-12.html">S</a>&nbsp;<a href="index-13.html">T</a>&nbsp;<a href="index-14.html">U</a>&nbsp;<a href="index-15.html">V</a>&nbsp;<a href="index-16.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a></main>
+<a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">E</a>&nbsp;<a href="index-6.html">F</a>&nbsp;<a href="index-7.html">G</a>&nbsp;<a href="index-8.html">H</a>&nbsp;<a href="index-9.html">L</a>&nbsp;<a href="index-10.html">M</a>&nbsp;<a href="index-11.html">P</a>&nbsp;<a href="index-12.html">R</a>&nbsp;<a href="index-13.html">S</a>&nbsp;<a href="index-14.html">T</a>&nbsp;<a href="index-15.html">U</a>&nbsp;<a href="index-16.html">V</a>&nbsp;<a href="index-17.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a></main>
 </div>
 </div>
 </body>
