+++ docs/index-files/index-11.html
@@ -1,12 +1,12 @@
 <!DOCTYPE HTML>
 <html lang="en">
 <head>
-<!-- Generated by javadoc (17) on Sun Dec 10 09:26:44 GMT 2023 -->
-<title>R-Index</title>
+<!-- Generated by javadoc (17) on Sun Dec 10 16:29:36 GMT 2023 -->
+<title>P-Index</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2023-12-10">
-<meta name="description" content="index: R">
+<meta name="description" content="index: P">
 <meta name="generator" content="javadoc/IndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.min.css" title="Style">
@@ -50,17 +50,27 @@ loadScripts(document, 'script');</script>
 <div class="header">
 <h1>Index</h1>
 </div>
-<a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">F</a>&nbsp;<a href="index-6.html">G</a>&nbsp;<a href="index-7.html">H</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">P</a>&nbsp;<a href="index-11.html">R</a>&nbsp;<a href="index-12.html">S</a>&nbsp;<a href="index-13.html">T</a>&nbsp;<a href="index-14.html">U</a>&nbsp;<a href="index-15.html">V</a>&nbsp;<a href="index-16.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a>
-<h2 class="title" id="I:R">R</h2>
+<a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">E</a>&nbsp;<a href="index-6.html">F</a>&nbsp;<a href="index-7.html">G</a>&nbsp;<a href="index-8.html">H</a>&nbsp;<a href="index-9.html">L</a>&nbsp;<a href="index-10.html">M</a>&nbsp;<a href="index-11.html">P</a>&nbsp;<a href="index-12.html">R</a>&nbsp;<a href="index-13.html">S</a>&nbsp;<a href="index-14.html">T</a>&nbsp;<a href="index-15.html">U</a>&nbsp;<a href="index-16.html">V</a>&nbsp;<a href="index-17.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a>
+<h2 class="title" id="I:P">P</h2>
 <dl class="index">
-<dt><a href="../com/github/westsi/vsi/Engine.html#Relation(java.util.HashMap,java.util.HashMap)" class="member-name-link">Relation(HashMap&lt;String, Integer&gt;, HashMap&lt;String, Integer&gt;)</a> - Method in class com.github.westsi.vsi.<a href="../com/github/westsi/vsi/Engine.html" title="class in com.github.westsi.vsi">Engine</a></dt>
+<dt><a href="../com/github/westsi/browser/BrowserTab.html#paintComponent(java.awt.Graphics)" class="member-name-link">paintComponent(Graphics)</a> - Method in class com.github.westsi.browser.<a href="../com/github/westsi/browser/BrowserTab.html" title="class in com.github.westsi.browser">BrowserTab</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../com/github/westsi/browser/util/url/URL.html#Request()" class="member-name-link">Request()</a> - Method in class com.github.westsi.browser.util.url.<a href="../com/github/westsi/browser/util/url/URL.html" title="class in com.github.westsi.browser.util.url">URL</a></dt>
+<dt><a href="../com/github/westsi/browser/util/Pair.html" class="type-name-link" title="class in com.github.westsi.browser.util">Pair</a>&lt;<a href="../com/github/westsi/browser/util/Pair.html" title="type parameter in Pair">F</a>,<wbr><a href="../com/github/westsi/browser/util/Pair.html" title="type parameter in Pair">S</a>&gt; - Class in <a href="../com/github/westsi/browser/util/package-summary.html">com.github.westsi.browser.util</a></dt>
+<dd>
+<div class="block">Utility class to group together two pieces of data.</div>
+</dd>
+<dt><a href="../com/github/westsi/browser/util/Pair.html#%3Cinit%3E(F,S)" class="member-name-link">Pair(F, S)</a> - Constructor for class com.github.westsi.browser.util.<a href="../com/github/westsi/browser/util/Pair.html" title="class in com.github.westsi.browser.util">Pair</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../com/github/westsi/vsi/Engine.html#RunEngine()" class="member-name-link">RunEngine()</a> - Static method in class com.github.westsi.vsi.<a href="../com/github/westsi/vsi/Engine.html" title="class in com.github.westsi.vsi">Engine</a></dt>
+<dt><a href="../com/github/westsi/browser/util/html/HTMLParser.html#parse(java.lang.String)" class="member-name-link">parse(String)</a> - Method in class com.github.westsi.browser.util.html.<a href="../com/github/westsi/browser/util/html/HTMLParser.html" title="class in com.github.westsi.browser.util.html">HTMLParser</a></dt>
+<dd>
+<div class="block">Parses the HTML source to the DOM Tree.</div>
+</dd>
+<dt><a href="../com/github/westsi/browser/util/html/HTMLParser.html#printTree(com.github.westsi.browser.util.html.HTMLElement,int)" class="member-name-link">printTree(HTMLElement, int)</a> - Method in class com.github.westsi.browser.util.html.<a href="../com/github/westsi/browser/util/html/HTMLParser.html" title="class in com.github.westsi.browser.util.html">HTMLParser</a></dt>
+<dd>&nbsp;</dd>
+<dt><a href="../com/github/westsi/browser/util/url/URLType.html#protocol" class="member-name-link">protocol</a> - Variable in enum class com.github.westsi.browser.util.url.<a href="../com/github/westsi/browser/util/url/URLType.html" title="enum class in com.github.westsi.browser.util.url">URLType</a></dt>
 <dd>&nbsp;</dd>
 </dl>
-<a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">F</a>&nbsp;<a href="index-6.html">G</a>&nbsp;<a href="index-7.html">H</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">P</a>&nbsp;<a href="index-11.html">R</a>&nbsp;<a href="index-12.html">S</a>&nbsp;<a href="index-13.html">T</a>&nbsp;<a href="index-14.html">U</a>&nbsp;<a href="index-15.html">V</a>&nbsp;<a href="index-16.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a></main>
+<a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">E</a>&nbsp;<a href="index-6.html">F</a>&nbsp;<a href="index-7.html">G</a>&nbsp;<a href="index-8.html">H</a>&nbsp;<a href="index-9.html">L</a>&nbsp;<a href="index-10.html">M</a>&nbsp;<a href="index-11.html">P</a>&nbsp;<a href="index-12.html">R</a>&nbsp;<a href="index-13.html">S</a>&nbsp;<a href="index-14.html">T</a>&nbsp;<a href="index-15.html">U</a>&nbsp;<a href="index-16.html">V</a>&nbsp;<a href="index-17.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a></main>
 </div>
 </div>
 </body>
