+++ docs/index-files/index-15.html
@@ -1,12 +1,12 @@
 <!DOCTYPE HTML>
 <html lang="en">
 <head>
-<!-- Generated by javadoc (17) on Sun Dec 10 09:26:44 GMT 2023 -->
-<title>V-Index</title>
+<!-- Generated by javadoc (17) on Sun Dec 10 16:29:36 GMT 2023 -->
+<title>U-Index</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2023-12-10">
-<meta name="description" content="index: V">
+<meta name="description" content="index: U">
 <meta name="generator" content="javadoc/IndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.min.css" title="Style">
@@ -50,24 +50,25 @@ loadScripts(document, 'script');</script>
 <div class="header">
 <h1>Index</h1>
 </div>
-<a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">F</a>&nbsp;<a href="index-6.html">G</a>&nbsp;<a href="index-7.html">H</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">P</a>&nbsp;<a href="index-11.html">R</a>&nbsp;<a href="index-12.html">S</a>&nbsp;<a href="index-13.html">T</a>&nbsp;<a href="index-14.html">U</a>&nbsp;<a href="index-15.html">V</a>&nbsp;<a href="index-16.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a>
-<h2 class="title" id="I:V">V</h2>
+<a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">E</a>&nbsp;<a href="index-6.html">F</a>&nbsp;<a href="index-7.html">G</a>&nbsp;<a href="index-8.html">H</a>&nbsp;<a href="index-9.html">L</a>&nbsp;<a href="index-10.html">M</a>&nbsp;<a href="index-11.html">P</a>&nbsp;<a href="index-12.html">R</a>&nbsp;<a href="index-13.html">S</a>&nbsp;<a href="index-14.html">T</a>&nbsp;<a href="index-15.html">U</a>&nbsp;<a href="index-16.html">V</a>&nbsp;<a href="index-17.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a>
+<h2 class="title" id="I:U">U</h2>
 <dl class="index">
-<dt><a href="../com/github/westsi/browser/util/url/URLType.html#valueOf(java.lang.String)" class="member-name-link">valueOf(String)</a> - Static method in enum class com.github.westsi.browser.util.url.<a href="../com/github/westsi/browser/util/url/URLType.html" title="enum class in com.github.westsi.browser.util.url">URLType</a></dt>
-<dd>
-<div class="block">Returns the enum constant of this class with the specified name.</div>
-</dd>
-<dt><a href="../com/github/westsi/browser/util/url/URLType.html#values()" class="member-name-link">values()</a> - Static method in enum class com.github.westsi.browser.util.url.<a href="../com/github/westsi/browser/util/url/URLType.html" title="enum class in com.github.westsi.browser.util.url">URLType</a></dt>
-<dd>
-<div class="block">Returns an array containing the constants of this enum class, in
-the order they are declared.</div>
-</dd>
-<dt><a href="../com/github/westsi/browser/util/url/URLType.html#VIEW_SOURCE" class="member-name-link">VIEW_SOURCE</a> - Enum constant in enum class com.github.westsi.browser.util.url.<a href="../com/github/westsi/browser/util/url/URLType.html" title="enum class in com.github.westsi.browser.util.url">URLType</a></dt>
+<dt><a href="../com/github/westsi/browser/BrowserTab.html#updateResize()" class="member-name-link">updateResize()</a> - Method in class com.github.westsi.browser.<a href="../com/github/westsi/browser/BrowserTab.html" title="class in com.github.westsi.browser">BrowserTab</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../com/github/westsi/browser/BrowserTab.html#VSTEP" class="member-name-link">VSTEP</a> - Static variable in class com.github.westsi.browser.<a href="../com/github/westsi/browser/BrowserTab.html" title="class in com.github.westsi.browser">BrowserTab</a></dt>
+<dt><a href="../com/github/westsi/browser/Layout.html#updateResize(java.lang.Integer,java.lang.Integer)" class="member-name-link">updateResize(Integer, Integer)</a> - Method in class com.github.westsi.browser.<a href="../com/github/westsi/browser/Layout.html" title="class in com.github.westsi.browser">Layout</a></dt>
 <dd>&nbsp;</dd>
+<dt><a href="../com/github/westsi/browser/util/url/URL.html" class="type-name-link" title="class in com.github.westsi.browser.util.url">URL</a> - Class in <a href="../com/github/westsi/browser/util/url/package-summary.html">com.github.westsi.browser.util.url</a></dt>
+<dd>
+<div class="block">Represents a URL and its type as specified in <code>URLType</code>.</div>
+</dd>
+<dt><a href="../com/github/westsi/browser/util/url/URL.html#%3Cinit%3E(java.lang.String)" class="member-name-link">URL(String)</a> - Constructor for class com.github.westsi.browser.util.url.<a href="../com/github/westsi/browser/util/url/URL.html" title="class in com.github.westsi.browser.util.url">URL</a></dt>
+<dd>&nbsp;</dd>
+<dt><a href="../com/github/westsi/browser/util/url/URLType.html" class="type-name-link" title="enum class in com.github.westsi.browser.util.url">URLType</a> - Enum Class in <a href="../com/github/westsi/browser/util/url/package-summary.html">com.github.westsi.browser.util.url</a></dt>
+<dd>
+<div class="block">The type of a URL that can be handled by the browser.</div>
+</dd>
 </dl>
-<a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">F</a>&nbsp;<a href="index-6.html">G</a>&nbsp;<a href="index-7.html">H</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">P</a>&nbsp;<a href="index-11.html">R</a>&nbsp;<a href="index-12.html">S</a>&nbsp;<a href="index-13.html">T</a>&nbsp;<a href="index-14.html">U</a>&nbsp;<a href="index-15.html">V</a>&nbsp;<a href="index-16.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a></main>
+<a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">E</a>&nbsp;<a href="index-6.html">F</a>&nbsp;<a href="index-7.html">G</a>&nbsp;<a href="index-8.html">H</a>&nbsp;<a href="index-9.html">L</a>&nbsp;<a href="index-10.html">M</a>&nbsp;<a href="index-11.html">P</a>&nbsp;<a href="index-12.html">R</a>&nbsp;<a href="index-13.html">S</a>&nbsp;<a href="index-14.html">T</a>&nbsp;<a href="index-15.html">U</a>&nbsp;<a href="index-16.html">V</a>&nbsp;<a href="index-17.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a></main>
 </div>
 </div>
 </body>
