+++ docs/index-files/index-9.html
@@ -1,12 +1,12 @@
 <!DOCTYPE HTML>
 <html lang="en">
 <head>
-<!-- Generated by javadoc (17) on Sun Dec 10 09:26:44 GMT 2023 -->
-<title>M-Index</title>
+<!-- Generated by javadoc (17) on Sun Dec 10 16:29:36 GMT 2023 -->
+<title>L-Index</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2023-12-10">
-<meta name="description" content="index: M">
+<meta name="description" content="index: L">
 <meta name="generator" content="javadoc/IndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.min.css" title="Style">
@@ -50,19 +50,27 @@ loadScripts(document, 'script');</script>
 <div class="header">
 <h1>Index</h1>
 </div>
-<a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">F</a>&nbsp;<a href="index-6.html">G</a>&nbsp;<a href="index-7.html">H</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">P</a>&nbsp;<a href="index-11.html">R</a>&nbsp;<a href="index-12.html">S</a>&nbsp;<a href="index-13.html">T</a>&nbsp;<a href="index-14.html">U</a>&nbsp;<a href="index-15.html">V</a>&nbsp;<a href="index-16.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a>
-<h2 class="title" id="I:M">M</h2>
+<a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">E</a>&nbsp;<a href="index-6.html">F</a>&nbsp;<a href="index-7.html">G</a>&nbsp;<a href="index-8.html">H</a>&nbsp;<a href="index-9.html">L</a>&nbsp;<a href="index-10.html">M</a>&nbsp;<a href="index-11.html">P</a>&nbsp;<a href="index-12.html">R</a>&nbsp;<a href="index-13.html">S</a>&nbsp;<a href="index-14.html">T</a>&nbsp;<a href="index-15.html">U</a>&nbsp;<a href="index-16.html">V</a>&nbsp;<a href="index-17.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a>
+<h2 class="title" id="I:L">L</h2>
 <dl class="index">
-<dt><a href="../com/github/westsi/vsi/Engine.html#Magnitude(java.util.HashMap)" class="member-name-link">Magnitude(HashMap&lt;String, Integer&gt;)</a> - Method in class com.github.westsi.vsi.<a href="../com/github/westsi/vsi/Engine.html" title="class in com.github.westsi.vsi">Engine</a></dt>
+<dt><a href="../com/github/westsi/browser/Layout.html#layout(com.github.westsi.browser.util.html.HTMLElement)" class="member-name-link">layout(HTMLElement)</a> - Method in class com.github.westsi.browser.<a href="../com/github/westsi/browser/Layout.html" title="class in com.github.westsi.browser">Layout</a></dt>
+<dd>
+<div class="block">Lays out tokens on page, with handling for some <code>HTMLTag</code>s and <code>HTMLText</code>.</div>
+</dd>
+<dt><a href="../com/github/westsi/browser/Layout.html" class="type-name-link" title="class in com.github.westsi.browser">Layout</a> - Class in <a href="../com/github/westsi/browser/package-summary.html">com.github.westsi.browser</a></dt>
+<dd>
+<div class="block">Lays out a web page for a <code>BrowserTab</code>.</div>
+</dd>
+<dt><a href="../com/github/westsi/browser/Layout.html#%3Cinit%3E(java.lang.Float,java.lang.Integer,java.lang.Integer)" class="member-name-link">Layout(Float, Integer, Integer)</a> - Constructor for class com.github.westsi.browser.<a href="../com/github/westsi/browser/Layout.html" title="class in com.github.westsi.browser">Layout</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../com/github/westsi/Main.html#main(java.lang.String%5B%5D)" class="member-name-link">main(String[])</a> - Static method in class com.github.westsi.<a href="../com/github/westsi/Main.html" title="class in com.github.westsi">Main</a></dt>
+<dt><a href="../com/github/westsi/browser/BrowserTab.html#LoadWebPage()" class="member-name-link">LoadWebPage()</a> - Method in class com.github.westsi.browser.<a href="../com/github/westsi/browser/BrowserTab.html" title="class in com.github.westsi.browser">BrowserTab</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../com/github/westsi/Main.html" class="type-name-link" title="class in com.github.westsi">Main</a> - Class in <a href="../com/github/westsi/package-summary.html">com.github.westsi</a></dt>
-<dd>&nbsp;</dd>
-<dt><a href="../com/github/westsi/Main.html#%3Cinit%3E()" class="member-name-link">Main()</a> - Constructor for class com.github.westsi.<a href="../com/github/westsi/Main.html" title="class in com.github.westsi">Main</a></dt>
-<dd>&nbsp;</dd>
+<dt><a href="../com/github/westsi/browser/Browser.html#LoadWebPage(java.lang.String)" class="member-name-link">LoadWebPage(String)</a> - Method in class com.github.westsi.browser.<a href="../com/github/westsi/browser/Browser.html" title="class in com.github.westsi.browser">Browser</a></dt>
+<dd>
+<div class="block">Creates a new <code>BrowserTab</code> and a <code>JScrollPane</code> to hold it, then loads the page in the new tab.</div>
+</dd>
 </dl>
-<a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">F</a>&nbsp;<a href="index-6.html">G</a>&nbsp;<a href="index-7.html">H</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">P</a>&nbsp;<a href="index-11.html">R</a>&nbsp;<a href="index-12.html">S</a>&nbsp;<a href="index-13.html">T</a>&nbsp;<a href="index-14.html">U</a>&nbsp;<a href="index-15.html">V</a>&nbsp;<a href="index-16.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a></main>
+<a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">E</a>&nbsp;<a href="index-6.html">F</a>&nbsp;<a href="index-7.html">G</a>&nbsp;<a href="index-8.html">H</a>&nbsp;<a href="index-9.html">L</a>&nbsp;<a href="index-10.html">M</a>&nbsp;<a href="index-11.html">P</a>&nbsp;<a href="index-12.html">R</a>&nbsp;<a href="index-13.html">S</a>&nbsp;<a href="index-14.html">T</a>&nbsp;<a href="index-15.html">U</a>&nbsp;<a href="index-16.html">V</a>&nbsp;<a href="index-17.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="../serialized-form.html">Serialized&nbsp;Form</a></main>
 </div>
 </div>
 </body>
