+++ docs/member-search-index.js
@@ -1,1 +1,1 @@
-memberSearchIndex = [{"p":"com.github.westsi.browser","c":"BrowserTab","l":"BrowserTab(String, Integer, Integer)","u":"%3Cinit%3E(java.lang.String,java.lang.Integer,java.lang.Integer)"},{"p":"com.github.westsi.vsi","c":"Engine","l":"Concordance(String)","u":"Concordance(java.lang.String)"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"DATA"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"FILE"},{"p":"com.github.westsi.browser.util","c":"StyledString","l":"font"},{"p":"com.github.westsi.browser","c":"Browser","l":"fonts"},{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"getAttributes()"},{"p":"com.github.westsi.browser.util.url","c":"URL","l":"GetDataUrlContents()"},{"p":"com.github.westsi.browser","c":"Layout","l":"getDisplayList()"},{"p":"com.github.westsi.vsi","c":"Engine","l":"getEngine()"},{"p":"com.github.westsi.browser.util.url","c":"URL","l":"GetFileContents()"},{"p":"com.github.westsi.browser.util","c":"Pair","l":"getFirst()"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"getFirst()"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"getFromURLPrefix(String)","u":"getFromURLPrefix(java.lang.String)"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"getGraphics()"},{"p":"com.github.westsi.browser","c":"Browser","l":"getInstance()"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"getName()"},{"p":"com.github.westsi.browser.util","c":"Pair","l":"getSecond()"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"getSecond()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"getTag()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLText","l":"getText()"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"getThird()"},{"p":"com.github.westsi.browser.util.url","c":"URL","l":"getUrlType()"},{"p":"com.github.westsi.browser","c":"Browser","l":"HEIGHT"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"HSTEP"},{"p":"com.github.westsi.browser.util.html","c":"HTMLParser","l":"HTMLParser()","u":"%3Cinit%3E()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"HTMLTag(String)","u":"%3Cinit%3E(java.lang.String)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLText","l":"HTMLText(String)","u":"%3Cinit%3E(java.lang.String)"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"HTTP"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"HTTPS"},{"p":"com.github.westsi.browser","c":"Layout","l":"Layout(Float, Integer, Integer)","u":"%3Cinit%3E(java.lang.Float,java.lang.Integer,java.lang.Integer)"},{"p":"com.github.westsi.browser","c":"Layout","l":"LayoutWebPage(ArrayList<HTMLElement>)","u":"LayoutWebPage(java.util.ArrayList)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLParser","l":"lex(String)","u":"lex(java.lang.String)"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"LoadWebPage()"},{"p":"com.github.westsi.browser","c":"Browser","l":"LoadWebPage(String)","u":"LoadWebPage(java.lang.String)"},{"p":"com.github.westsi.vsi","c":"Engine","l":"Magnitude(HashMap<String, Integer>)","u":"Magnitude(java.util.HashMap)"},{"p":"com.github.westsi","c":"Main","l":"Main()","u":"%3Cinit%3E()"},{"p":"com.github.westsi","c":"Main","l":"main(String[])","u":"main(java.lang.String[])"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"paintComponent(Graphics)","u":"paintComponent(java.awt.Graphics)"},{"p":"com.github.westsi.browser.util","c":"Pair","l":"Pair(F, S)","u":"%3Cinit%3E(F,S)"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"protocol"},{"p":"com.github.westsi.vsi","c":"Engine","l":"Relation(HashMap<String, Integer>, HashMap<String, Integer>)","u":"Relation(java.util.HashMap,java.util.HashMap)"},{"p":"com.github.westsi.browser.util.url","c":"URL","l":"Request()"},{"p":"com.github.westsi.vsi","c":"Engine","l":"RunEngine()"},{"p":"com.github.westsi.browser.util","c":"Pair","l":"setFirst(F)"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"setFirst(F)"},{"p":"com.github.westsi.browser.util","c":"Pair","l":"setSecond(S)"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"setSecond(S)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLText","l":"setText(String)","u":"setText(java.lang.String)"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"setThird(T)"},{"p":"com.github.westsi.browser.util","c":"StyledString","l":"string"},{"p":"com.github.westsi.browser.util","c":"StyledString","l":"StyledString(String, Font)","u":"%3Cinit%3E(java.lang.String,java.awt.Font)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLElement","l":"toString()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"toString()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLText","l":"toString()"},{"p":"com.github.westsi.browser.util","c":"Pair","l":"toString()"},{"p":"com.github.westsi.browser.util","c":"StyledString","l":"toString()"},{"p":"com.github.westsi.browser.util.url","c":"URL","l":"toString()"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"Triplet(F, S, T)","u":"%3Cinit%3E(F,S,T)"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"updateResize()"},{"p":"com.github.westsi.browser","c":"Layout","l":"updateResize(Integer, Integer)","u":"updateResize(java.lang.Integer,java.lang.Integer)"},{"p":"com.github.westsi.browser.util.url","c":"URL","l":"URL(String)","u":"%3Cinit%3E(java.lang.String)"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"valueOf(String)","u":"valueOf(java.lang.String)"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"values()"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"VIEW_SOURCE"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"VSTEP"},{"p":"com.github.westsi.browser","c":"Browser","l":"WIDTH"}];updateSearchResults();
+memberSearchIndex = [{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"addAttribute(String, String)","u":"addAttribute(java.lang.String,java.lang.String)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLElement","l":"addChild(HTMLElement)","u":"addChild(com.github.westsi.browser.util.html.HTMLElement)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"addChild(HTMLElement)","u":"addChild(com.github.westsi.browser.util.html.HTMLElement)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLText","l":"addChild(HTMLElement)","u":"addChild(com.github.westsi.browser.util.html.HTMLElement)"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"BrowserTab(String, Integer, Integer)","u":"%3Cinit%3E(java.lang.String,java.lang.Integer,java.lang.Integer)"},{"p":"com.github.westsi.vsi","c":"Engine","l":"Concordance(String)","u":"Concordance(java.lang.String)"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"DATA"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"FILE"},{"p":"com.github.westsi.browser.util","c":"StyledString","l":"font"},{"p":"com.github.westsi.browser","c":"Browser","l":"fonts"},{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"getAttributes()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLElement","l":"getChildren()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"getChildren()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLText","l":"getChildren()"},{"p":"com.github.westsi.browser.util.url","c":"URL","l":"GetDataUrlContents()"},{"p":"com.github.westsi.browser","c":"Layout","l":"getDisplayList()"},{"p":"com.github.westsi.vsi","c":"Engine","l":"getEngine()"},{"p":"com.github.westsi.browser.util.url","c":"URL","l":"GetFileContents()"},{"p":"com.github.westsi.browser.util","c":"Pair","l":"getFirst()"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"getFirst()"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"getFromURLPrefix(String)","u":"getFromURLPrefix(java.lang.String)"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"getGraphics()"},{"p":"com.github.westsi.browser","c":"Browser","l":"getInstance()"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"getName()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLElement","l":"getParent()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"getParent()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLText","l":"getParent()"},{"p":"com.github.westsi.browser.util","c":"Pair","l":"getSecond()"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"getSecond()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"getTag()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLText","l":"getText()"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"getThird()"},{"p":"com.github.westsi.browser.util.url","c":"URL","l":"getUrlType()"},{"p":"com.github.westsi.browser","c":"Browser","l":"HEIGHT"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"HSTEP"},{"p":"com.github.westsi.browser.util.html","c":"HTMLParser","l":"HTMLParser()","u":"%3Cinit%3E()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"HTMLTag(String, HTMLElement)","u":"%3Cinit%3E(java.lang.String,com.github.westsi.browser.util.html.HTMLElement)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"HTMLTag(String, HTMLElement, HashMap<String, String>)","u":"%3Cinit%3E(java.lang.String,com.github.westsi.browser.util.html.HTMLElement,java.util.HashMap)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLText","l":"HTMLText(String, HTMLElement)","u":"%3Cinit%3E(java.lang.String,com.github.westsi.browser.util.html.HTMLElement)"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"HTTP"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"HTTPS"},{"p":"com.github.westsi.browser","c":"Layout","l":"Layout(Float, Integer, Integer)","u":"%3Cinit%3E(java.lang.Float,java.lang.Integer,java.lang.Integer)"},{"p":"com.github.westsi.browser","c":"Layout","l":"layout(HTMLElement)","u":"layout(com.github.westsi.browser.util.html.HTMLElement)"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"LoadWebPage()"},{"p":"com.github.westsi.browser","c":"Browser","l":"LoadWebPage(String)","u":"LoadWebPage(java.lang.String)"},{"p":"com.github.westsi.vsi","c":"Engine","l":"Magnitude(HashMap<String, Integer>)","u":"Magnitude(java.util.HashMap)"},{"p":"com.github.westsi","c":"Main","l":"Main()","u":"%3Cinit%3E()"},{"p":"com.github.westsi","c":"Main","l":"main(String[])","u":"main(java.lang.String[])"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"paintComponent(Graphics)","u":"paintComponent(java.awt.Graphics)"},{"p":"com.github.westsi.browser.util","c":"Pair","l":"Pair(F, S)","u":"%3Cinit%3E(F,S)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLParser","l":"parse(String)","u":"parse(java.lang.String)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLParser","l":"printTree(HTMLElement, int)","u":"printTree(com.github.westsi.browser.util.html.HTMLElement,int)"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"protocol"},{"p":"com.github.westsi.vsi","c":"Engine","l":"Relation(HashMap<String, Integer>, HashMap<String, Integer>)","u":"Relation(java.util.HashMap,java.util.HashMap)"},{"p":"com.github.westsi.browser.util.url","c":"URL","l":"Request()"},{"p":"com.github.westsi.browser","c":"Layout","l":"resetVariables()"},{"p":"com.github.westsi.vsi","c":"Engine","l":"RunEngine()"},{"p":"com.github.westsi.browser.util","c":"Pair","l":"setFirst(F)"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"setFirst(F)"},{"p":"com.github.westsi.browser.util","c":"Pair","l":"setSecond(S)"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"setSecond(S)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLText","l":"setText(String)","u":"setText(java.lang.String)"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"setThird(T)"},{"p":"com.github.westsi.browser.util","c":"StyledString","l":"string"},{"p":"com.github.westsi.browser.util","c":"StyledString","l":"StyledString(String, Font)","u":"%3Cinit%3E(java.lang.String,java.awt.Font)"},{"p":"com.github.westsi.browser.util.html","c":"HTMLElement","l":"toString()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLTag","l":"toString()"},{"p":"com.github.westsi.browser.util.html","c":"HTMLText","l":"toString()"},{"p":"com.github.westsi.browser.util","c":"Pair","l":"toString()"},{"p":"com.github.westsi.browser.util","c":"StyledString","l":"toString()"},{"p":"com.github.westsi.browser.util.url","c":"URL","l":"toString()"},{"p":"com.github.westsi.browser.util","c":"Triplet","l":"Triplet(F, S, T)","u":"%3Cinit%3E(F,S,T)"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"updateResize()"},{"p":"com.github.westsi.browser","c":"Layout","l":"updateResize(Integer, Integer)","u":"updateResize(java.lang.Integer,java.lang.Integer)"},{"p":"com.github.westsi.browser.util.url","c":"URL","l":"URL(String)","u":"%3Cinit%3E(java.lang.String)"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"valueOf(String)","u":"valueOf(java.lang.String)"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"values()"},{"p":"com.github.westsi.browser.util.url","c":"URLType","l":"VIEW_SOURCE"},{"p":"com.github.westsi.browser","c":"BrowserTab","l":"VSTEP"},{"p":"com.github.westsi.browser","c":"Browser","l":"WIDTH"}];updateSearchResults();
