+++ docs/serialized-form.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="en">
 <head>
-<!-- Generated by javadoc (17) on Sun Dec 10 09:26:44 GMT 2023 -->
+<!-- Generated by javadoc (17) on Sun Dec 10 16:29:36 GMT 2023 -->
 <title>Serialized Form</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
@@ -73,12 +73,12 @@ loadScripts(document, 'script');</script>
 <pre><a href="com/github/westsi/browser/Layout.html" title="class in com.github.westsi.browser">Layout</a> layout</pre>
 </li>
 <li class="block-list">
-<h5>scrolled</h5>
-<pre><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Integer.html" title="class or interface in java.lang" class="external-link">Integer</a> scrolled</pre>
+<h5>root</h5>
+<pre><a href="com/github/westsi/browser/util/html/HTMLElement.html" title="interface in com.github.westsi.browser.util.html">HTMLElement</a> root</pre>
 </li>
 <li class="block-list">
-<h5>tokens</h5>
-<pre><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/util/ArrayList.html" title="class or interface in java.util" class="external-link">ArrayList</a>&lt;<a href="com/github/westsi/browser/util/html/HTMLElement.html" title="interface in com.github.westsi.browser.util.html">HTMLElement</a>&gt; tokens</pre>
+<h5>scrolled</h5>
+<pre><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Integer.html" title="class or interface in java.lang" class="external-link">Integer</a> scrolled</pre>
 </li>
 <li class="block-list">
 <h5>url</h5>
