+++ src/main/java/com/github/westsi/Main.java
@@ -15,10 +15,10 @@ public class Main {
 //        String url = userInput.nextLine();
         Browser browser = Browser.getInstance();
 //        browser.LoadWebPage(url);
-        browser.LoadWebPage("https://gist.githubusercontent.com/Westsi/58b5aeeb2c53e6682bed13eacdb7be8c/raw/6317dacce0eb91f924e1d0ea43bf68ace889f89b/test.html");
+//        browser.LoadWebPage("https://gist.githubusercontent.com/Westsi/58b5aeeb2c53e6682bed13eacdb7be8c/raw/6317dacce0eb91f924e1d0ea43bf68ace889f89b/test.html");
 //        browser.LoadWebPage("https://browser.engineering/examples/example3-sizes.html");
-//        browser.LoadWebPage("https://browser.engineering/html.html");
-        browser.LoadWebPage("https://hweeo");
-        browser.LoadWebPage("data:text/html,<b>Hi&lt;&gt;</b>");
+        browser.LoadWebPage("https://browser.engineering/html.html");
+//        browser.LoadWebPage("https://hweeo");
+//        browser.LoadWebPage("data:text/html,<b>Hi&lt;&gt;</b>");
     }
 }
